youle-mvp/pyproject.toml
[project]
name = "youle-mvp"
version = "0.1.0"
description = "「有了 / Youle」AI Agent V0 反诈短视频 demo"
requires-python = ">=3.12,<3.13"
dependencies = [
    # === Web 框架 ===
    "fastapi==0.115.6",
    "uvicorn[standard]==0.32.1",
    "python-multipart==0.0.20",
    "python-dotenv==1.0.1",
    "aiofiles==24.1.0",
    "sse-starlette==2.1.3",
    "orjson==3.10.12",

    # === LangGraph 1.x ===
    "langgraph==1.1.9",
    "langgraph-checkpoint-sqlite==3.0.3",

    # === LLM SDK ===
    "anthropic==0.42.0",
    "httpx==0.27.2",

    # === 数据 + 文件 ===
    "pandas==2.2.3",
    "openpyxl==3.1.5",
    "pillow==11.0.0",
    "beautifulsoup4==4.12.3",

    # === 视频 ===
    "ffmpeg-python==0.2.0",

    # === 校验 ===
    "pydantic>=2.10.4",
    "pydantic-settings>=2.7.0",

    # === 可靠性 ===
    "tenacity==8.5.0",

    # === 日志 ===
    "structlog==24.4.0",
    "rich==13.7.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.25.0",
    "ruff==0.8.4",
    "mypy==1.13.0",
    "respx==0.21.1",
]
v1 = [
    "sqlalchemy[asyncio]==2.0.36",
    "asyncpg==0.29.0",
    "alembic==1.13.3",
    "redis==5.0.8",
    "arq==0.26.1",
    "litellm==1.50.4",
    "fastapi-users[sqlalchemy]==14.0.1",
    "slowapi==0.1.9",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
